move StatementExecuteUnit to sql package
